refactor: consolidate OBR integration flow and remove legacy modules
- Unify OBR helper utilities under `burundi_compliance.utils`, remove deprecated
`api_classes`, `data`, and old utility modules, and update all API/background
task/override imports to the new shared paths to simplify maintenance.

- Also update invoice submission workflows by introducing deferred OBR submission
handling (`custom_defer_submission_to_obr`), repoint POS bulk submission to the
new whitelisted API method, clean up hooks/client script wiring
