+++ burundi_compliance/burundi_compliance/custom/pos_invoice_item.json
@@ -5,9 +5,11 @@
    "_comments": null,
    "_liked_by": null,
    "_user_tags": null,
+   "alignment": null,
    "allow_in_quick_entry": 0,
    "allow_on_submit": 0,
    "bold": 0,
+   "button_color": null,
    "collapsible": 0,
    "collapsible_depends_on": null,
    "columns": 0,
@@ -37,6 +39,7 @@
    "is_virtual": 0,
    "label": "Vehicle",
    "length": 0,
+   "link_filters": null,
    "mandatory_depends_on": null,
    "modified": "2023-12-18 07:09:43.370709",
    "modified_by": "bnjogu@navari.co.ke",
@@ -47,6 +50,7 @@
    "options": "Vehicle",
    "owner": "bnjogu@navari.co.ke",
    "permlevel": 0,
+   "placeholder": null,
    "precision": "",
    "print_hide": 0,
    "print_hide_if_no_value": 0,
@@ -80,8 +84,8 @@
    "field_name": "barcode",
    "idx": 0,
    "is_system_generated": 1,
-   "modified": "2024-03-08 09:13:02.325833",
-   "modified_by": "bnjogu@navari.co.ke",
+   "modified": "2026-04-12 09:57:18.074419",
+   "modified_by": "Administrator",
    "module": null,
    "name": "POS Invoice Item-barcode-hidden",
    "owner": "bnjogu@navari.co.ke",
