--- burundi_compliance/burundi_compliance/data/test_data.py
@@ -1,59 +1,0 @@
-def prepare_test_invoice_data():
-    return {
-        "invoice_number": "INV001",
-        "invoice_date": "2021-12-06 07:30:22",
-        "invoice_type": "FN",
-        "tp_type": "1",
-        "tp_name": "NDIKUMANA JEAN MARIE",
-        "tp_TIN": "4400773244",
-        "tp_trade_number": "3333",
-        "tp_postal_number": "3256",
-        "tp_phone_number": "70959595",
-        "tp_address_province": "BUJUMBURA",
-        "tp_address_commune": "BUJUMBURA",
-        "tp_address_quartier": "GIKUNGU",
-        "tp_address_avenue": "MUYINGA",
-        "tp_address_number": "",
-        "vat_taxpayer": "1",
-        "ct_taxpayer": "0",
-        "tl_taxpayer": "0",
-        "tp_fiscal_center": "DGC",
-        "tp_activity_sector": "SERVICE MARCHAND",
-        "tp_legal_form": "suprl",
-        "payment_type": "1",
-        "invoice_currency": "BIF",
-        "customer_name": "NGARUKIYINTWARI WAKA",
-        "customer_TIN": "4100022020",
-        "customer_address": "KIRUNDO",
-        "vat_customer_payer": "1",
-        "cancelled_invoice_ref": "",
-        "invoice_ref": "",
-        "cn_motif": "",
-        "invoice_identifier": "4400773244/ws440077324400027/20211206073022/0001/2021",
-        "invoice_items": [
-            {
-                "item_designation": "ARTICLE ONE",
-                "item_quantity": "10",
-                "item_price": "500",
-                "item_ct": "789",
-                "item_tl": "123",
-                "item_price_nvat": "5789",
-                "vat": "1042.02",
-                "item_price_wvat": "6831.02",
-                "item_total_amount": "6954.02",
-            },
-            {
-                "item_designation": "ARTICLE TWO",
-                "item_quantity": "10",
-                "item_price": "900",
-                "item_ct": "0",
-                "item_tl": "0",
-                "item_ott_tax": "0",
-                "item_tsce_tax": "0",
-                "item_price_nvat": "9000",
-                "vat": "1620",
-                "item_price_wvat": "10620",
-                "item_total_amount": "10620",
-            },
-        ]
-    }
